SIG-873 Split out v1 view, serializer, serializer fields code
